+++ CHANGELOG.md
@@ -1,3 +1,11 @@
+## 0.0.3-beta
+
+* Fixed issue in image_input while selecting image
+* Improved preview image screen
+## 0.0.2-beta
+
+* Updated image_picker export to show only XFile
+
 ## 0.0.1-beta
 
 * initial beta release of image_input